+++ lib/src/widgets/datetime_text_field.dart
@@ -1,0 +1,232 @@
+import 'dart:math';
+
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:intl/intl.dart' hide TextDirection;
+
+class DateTimeTextField extends StatelessWidget {
+  const DateTimeTextField({
+    super.key,
+    this.controller,
+    this.value,
+    this.format = 'dd/MM/yyyy',
+    this.showDateTimePicker,
+    this.focusNode,
+    this.decoration = const InputDecoration(),
+    this.keyboardType,
+    this.textInputAction,
+    this.textCapitalization = TextCapitalization.none,
+    this.style,
+    this.strutStyle,
+    this.textAlign = TextAlign.center,
+    this.textAlignVertical,
+    this.textDirection,
+    this.readOnly = false,
+    this.obscureText = false,
+    this.showCursor,
+    this.autofocus = false,
+    this.obscuringCharacter = '•',
+    this.autocorrect = true,
+    this.smartDashesType,
+    this.smartQuotesType,
+    this.enableSuggestions = true,
+    this.expands = false,
+    this.maxLength,
+    this.onChanged,
+    this.onEditingComplete,
+    this.onAppPrivateCommand,
+    this.enabled,
+    this.cursorWidth = 2.0,
+    this.cursorHeight,
+    this.cursorRadius,
+    this.cursorOpacityAnimates,
+    this.cursorColor,
+    this.keyboardAppearance,
+    this.scrollPadding = const EdgeInsets.all(20.0),
+    this.dragStartBehavior = DragStartBehavior.start,
+    this.selectionControls,
+    this.onTap,
+    this.onTapOutside,
+    this.mouseCursor,
+    this.buildCounter,
+    this.scrollController,
+    this.scrollPhysics,
+    this.autofillHints = const <String>[],
+    this.clipBehavior = Clip.hardEdge,
+    this.restorationId,
+    this.scribbleEnabled = true,
+    this.enableIMEPersonalizedLearning = true,
+    this.canRequestFocus = true,
+    this.magnifierConfiguration,
+    this.enableInteractiveSelection,
+  });
+
+  final DateTime? value;
+  final TextEditingController? controller;
+
+  /// Formato fo the date
+  ///
+  /// default is dd/MM/yyyy
+  final String format;
+
+  final Future<DateTime?> Function(BuildContext context, DateTime? initialDate)?
+      showDateTimePicker;
+
+  final TextMagnifierConfiguration? magnifierConfiguration;
+  final FocusNode? focusNode;
+  final InputDecoration? decoration;
+  final TextInputType? keyboardType;
+  final TextInputAction? textInputAction;
+  final TextCapitalization textCapitalization;
+  final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign textAlign;
+  final TextAlignVertical? textAlignVertical;
+  final TextDirection? textDirection;
+  final bool autofocus;
+  final bool obscureText;
+  final String obscuringCharacter;
+  final bool autocorrect;
+  final SmartDashesType? smartDashesType;
+  final SmartQuotesType? smartQuotesType;
+  final bool enableSuggestions;
+  final bool expands;
+  final bool readOnly;
+  final bool? showCursor;
+  final int? maxLength;
+  final ValueChanged<DateTime>? onChanged;
+  final VoidCallback? onEditingComplete;
+  final AppPrivateCommandCallback? onAppPrivateCommand;
+  final bool? enabled;
+  final double cursorWidth;
+  final double? cursorHeight;
+  final Radius? cursorRadius;
+  final bool? cursorOpacityAnimates;
+  final Color? cursorColor;
+  final Brightness? keyboardAppearance;
+  final EdgeInsets scrollPadding;
+  final bool? enableInteractiveSelection;
+  final TextSelectionControls? selectionControls;
+  final DragStartBehavior dragStartBehavior;
+  final GestureTapCallback? onTap;
+  final TapRegionCallback? onTapOutside;
+  final MouseCursor? mouseCursor;
+  final InputCounterWidgetBuilder? buildCounter;
+  final ScrollPhysics? scrollPhysics;
+  final ScrollController? scrollController;
+  final Iterable<String>? autofillHints;
+  final Clip clipBehavior;
+  final String? restorationId;
+  final bool scribbleEnabled;
+  final bool enableIMEPersonalizedLearning;
+  final bool canRequestFocus;
+
+  DateFormat get _format => DateFormat(format);
+
+  void selectDateTime(BuildContext context) async {
+    final date = showDateTimePicker == null
+        ? await showDatePicker(
+            // ignore: use_build_context_synchronously
+            context: context,
+            firstDate: DateTime(1970),
+            lastDate: DateTime(2100),
+            initialDate: value,
+          )
+        // ignore: use_build_context_synchronously
+        : await showDateTimePicker?.call(context, value);
+    if (date != null) {
+      onChanged?.call(date);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    TextEditingController fieldController =
+        (controller ?? TextEditingController());
+    if (value != null) {
+      fieldController.text = _format.format(value!);
+    }
+
+    return TextFormField(
+      controller: fieldController,
+      obscureText: obscureText,
+      inputFormatters: [DateInputFormatter()],
+      decoration: decoration,
+      keyboardType: keyboardType ?? TextInputType.datetime,
+      textInputAction: textInputAction,
+      textCapitalization: textCapitalization,
+      style: style,
+      strutStyle: strutStyle,
+      textAlign: textAlign,
+      textAlignVertical: textAlignVertical,
+      textDirection: textDirection,
+      readOnly: readOnly,
+      showCursor: showCursor,
+      autofocus: autofocus,
+      obscuringCharacter: obscuringCharacter,
+      autocorrect: autocorrect,
+      smartDashesType: smartDashesType,
+      smartQuotesType: smartQuotesType,
+      enableSuggestions: enableSuggestions,
+      expands: expands,
+      maxLength: maxLength,
+      // onChanged: onChanged,
+      onEditingComplete: onEditingComplete,
+      onAppPrivateCommand: onAppPrivateCommand,
+      enabled: enabled,
+      cursorWidth: cursorWidth,
+      cursorHeight: cursorHeight,
+      cursorRadius: cursorRadius,
+      cursorOpacityAnimates: cursorOpacityAnimates,
+      cursorColor: cursorColor,
+      keyboardAppearance: keyboardAppearance,
+      scrollPadding: scrollPadding,
+      dragStartBehavior: dragStartBehavior,
+      selectionControls: selectionControls,
+      onTap: () => selectDateTime(context),
+      onTapOutside: onTapOutside,
+      mouseCursor: mouseCursor,
+      buildCounter: buildCounter,
+      scrollController: scrollController,
+      scrollPhysics: scrollPhysics,
+      autofillHints: autofillHints,
+      clipBehavior: clipBehavior,
+      restorationId: restorationId,
+      scribbleEnabled: scribbleEnabled,
+      enableIMEPersonalizedLearning: enableIMEPersonalizedLearning,
+      canRequestFocus: canRequestFocus,
+      magnifierConfiguration: magnifierConfiguration,
+      enableInteractiveSelection: enableInteractiveSelection ?? false,
+    );
+  }
+}
+
+class DateInputFormatter extends TextInputFormatter {
+  static const _maxChars = 8;
+
+  @override
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
+    var text = _format(newValue.text, '/');
+    return newValue.copyWith(text: text, selection: updateCursorPosition(text));
+  }
+
+  String _format(String value, String seperator) {
+    value = value.replaceAll(seperator, '');
+    var newString = '';
+
+    for (int i = 0; i < min(value.length, _maxChars); i++) {
+      newString += value[i];
+      if ((i == 1 || i == 3) && i != value.length - 1) {
+        newString += seperator;
+      }
+    }
+
+    return newString;
+  }
+
+  TextSelection updateCursorPosition(String text) {
+    return TextSelection.fromPosition(TextPosition(offset: text.length));
+  }
+}