+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.5
+
+Adding new types of fields
+- DateTimeTextField for the DateTime
+- TimeOfDayTextField for the TimeOfDay
+
 ## 0.2.2
 
 Creating forms faster with Cubit or Flutter Bloc